+++ lib/src/eval/compiler/declaration/field.dart
@@ -6,7 +6,8 @@ import 'package:dart_eval/src/eval/compiler/scope.dart';
 import 'package:dart_eval/src/eval/compiler/type.dart';
 import 'package:dart_eval/src/eval/runtime/runtime.dart';
 
-void compileFieldDeclaration(int fieldIndex, FieldDeclaration d, CompilerContext ctx, ClassDeclaration parent) {
+void compileFieldDeclaration(
+    int fieldIndex, FieldDeclaration d, CompilerContext ctx, NamedCompilationUnitMember parent) {
   final parentName = parent.name.value() as String;
   var _fieldIndex = fieldIndex;
   for (final field in d.fields.variables) {