+++ lib/src/eval/compiler/helpers/tearoff.dart
@@ -135,6 +135,7 @@ extension TearOff on Variable {
     final requiredPositionalArgCount = positional.where((element) => element.isRequired).length;
 
     final positionalArgTypes = positional
+        .map((a) => a is NormalFormalParameter ? a : (a as DefaultFormalParameter).parameter)
         .cast<SimpleFormalParameter>()
         .map((a) => a.type == null ? EvalTypes.dynamicType : TypeRef.fromAnnotation(ctx, ctx.library, a.type!))
         .map((t) => t.toRuntimeType(ctx))